+++ CHANGES.md
@@ -1,2 +1,8 @@
-- Show optimization button for crafts without qualities but with quality reagents
 - Major refactoring of GUI code
+- Show optimization button for crafts without qualities but with quality reagents
+- Allow tracked recipe amounts to go to 0 intead of stopping at 1
+- Track and show result quality in recipe tracker
+- Load optimal allocations for tracked recipes on login
+- Show missing instead of owned reagent count for missing reagents in reagents tracker
+- Show reagents that are outputs of tracked recipes in separate reagents tracker category
+- Handle clicks on crafted reagents in reagents tracker as clicks on the corresponding recipe